refactor(terminal): consolidate metrics test helpers into test_helpers
DebuggingRecorder install + counter_value + histogram_count +
histogram_samples were duplicated across observability_metrics.rs,
jk_scroll_synthetic.rs, and pty/reader.rs unit tests. Hoisted to
service::test_helpers under the existing test-helpers feature gate;
three call sites now share one canonical implementation.

metrics-util moves from a pure dev-dependency to an optional
dependency activated by the test-helpers feature so the shared
helpers are visible to integration tests via the public path
`ozmux_terminal::service::test_helpers::*`.
